+++ esp32/lib/module_handler/include/modules/wifi.hpp
@@ -80,6 +80,12 @@ class ModuleWiFi : public ModuleHandler::Module {
 
   void NtpSync(const Esp32Command &cmd);
 
+  void Host(const Esp32Command &cmd);
+
+  void StopHost();
+
+  void HostInfo();
+
   WiFiUDP ntpUDP;
 
   NTPClient *timeClient;