+++ CHANGELOG.md
@@ -8,12 +8,32 @@ and this project adheres to
 minor versions follow hardware versioning with patches used to iterate on
 firmware versions for a hardware version.
 
-## [Unreleased] - TBD
+## [Unreleased]
+
+### Updated
+
+- Esp32 enters a deep sleep state when idle [#150](i150)
+- Upload interval for LoRaWAN
+
+### Fixed
+
+- Updated clock configuration [#268](i268)
+
+[i150]: https://github.com/jlab-sensing/ENTS-node-firmware/issues/150
+[i268]: https://github.com/jlab-sensing/ENTS-node-firmware/issues/268
+
+## [2.3.4] - 2025-08-20
 
 ### Added
 
 - Example to force esp32 in deep sleep [#207](i207)
 - Utility to change SDI-12 sensor address [#210](i210)
+- Added new sensors
+    - SEN0308
+    - SEN0257
+    - YFS210C
+- Added controlling of solenoid from esp32 api
+- Add user configuration vai WiFi [#259](i259)
 
 ### Updated
 
@@ -27,6 +47,7 @@ firmware versions for a hardware version.
 [i210]: https://github.com/jlab-sensing/ENTS-node-firmware/issues/210
 [i252]: https://github.com/jlab-sensing/ENTS-node-firmware/issues/252
 [pr249]: https://github.com/jlab-sensing/ENTS-node-firmware/pull/249
+[i259]: https://github.com/jlab-sensing/ENTS-node-firmware/issues/259
 
 ## [2.3.3] - 2025-06-18